pyproject.toml
[tool.poetry]
name = "bento-chat-assistant"
version = "0.1.0"
description = "FastAPI based Chatbot Starter Kit"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
pywebview = "^4.0.0"
anthropic = "^0.18.1"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
